requirements.txt
langchain
openai>=1.0.0
faiss-cpu
sentence-transformers
pandas
scikit-learn
pinecone>=5.0.0
python-dotenv